Move root loading state to root route file
