--- client/src/components/RootLoadingState.module.scss
@@ -1,8 +1,0 @@
-.header {
-  justify-content: end;
-  z-index: 1;
-}
-
-.skeleton__playlist {
-  padding: 0.5rem 0;
-}
